Worker function used in extractSummaries_1section

Worker function used in extractSummaries_1section

Usage

extractSummaries_1plan(
  arglist,
  sectionHeaders,
  sectionFields,
  textToParse,
  filename
)

Arguments

arglist

The argument list

sectionHeaders

A character vector with headers for each section of interest

sectionFields

is a list of data.frames where each data.frame specifies the fields to be extracted for that section

textToParse

The text to parse

filename

The filename

Value

A list

Examples

# make me!!!