Load an analysis dataset from the SAVEDATA command into an R data.frame

This function reads an analysis dataset generated by the Mplus SAVEDATA command and returns an R data.frame object.

Usage

getSavedata_Data(outfile)

Arguments

outfile

Required. The name of the Mplus output file to read. Can be an absolute or relative path. If outfile is a relative path or just the filename, then it is assumed that the file resides in the working directory getwd().

Value

A data.frame containing the analysis dataset generated by the SAVEDATA command.

Note

Note that the outfile parameter should refer to the Mplus output file (.out extension), not the actual dataset generated by SAVEDATA. This function reads information about the dataset from the .out file and loads the dataset accordingly.

See also

Author

Michael Hallquist

Examples

if (FALSE) { # \dontrun{
  savedat <- getSavedata_Data("C:/Program Files/Mplus/Test Output.out")
} # }