Split a data frame into a list by rows

Takes a data frame and returns a list with an element for each row of the data frame. This is an internal function.

Usage

splitDFByRow(df)

Arguments

df

An object inheriting from class data.frame

Value

A list where each element is a one row data frame

Examples

# small example using built in data
MplusAutomation:::splitDFByRow(mtcars)
#> [[1]]
#>           mpg cyl disp  hp drat   wt  qsec vs am gear carb
#> Mazda RX4  21   6  160 110  3.9 2.62 16.46  0  1    4    4
#> 
#> [[2]]
#>               mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4 Wag  21   6  160 110  3.9 2.875 17.02  0  1    4    4
#> 
#> [[3]]
#>             mpg cyl disp hp drat   wt  qsec vs am gear carb
#> Datsun 710 22.8   4  108 93 3.85 2.32 18.61  1  1    4    1
#> 
#> [[4]]
#>                 mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Hornet 4 Drive 21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> 
#> [[5]]
#>                    mpg cyl disp  hp drat   wt  qsec vs am gear carb
#> Hornet Sportabout 18.7   8  360 175 3.15 3.44 17.02  0  0    3    2
#> 
#> [[6]]
#>          mpg cyl disp  hp drat   wt  qsec vs am gear carb
#> Valiant 18.1   6  225 105 2.76 3.46 20.22  1  0    3    1
#> 
#> [[7]]
#>             mpg cyl disp  hp drat   wt  qsec vs am gear carb
#> Duster 360 14.3   8  360 245 3.21 3.57 15.84  0  0    3    4
#> 
#> [[8]]
#>            mpg cyl  disp hp drat   wt qsec vs am gear carb
#> Merc 240D 24.4   4 146.7 62 3.69 3.19   20  1  0    4    2
#> 
#> [[9]]
#>           mpg cyl  disp hp drat   wt qsec vs am gear carb
#> Merc 230 22.8   4 140.8 95 3.92 3.15 22.9  1  0    4    2
#> 
#> [[10]]
#>           mpg cyl  disp  hp drat   wt qsec vs am gear carb
#> Merc 280 19.2   6 167.6 123 3.92 3.44 18.3  1  0    4    4
#> 
#> [[11]]
#>            mpg cyl  disp  hp drat   wt qsec vs am gear carb
#> Merc 280C 17.8   6 167.6 123 3.92 3.44 18.9  1  0    4    4
#> 
#> [[12]]
#>             mpg cyl  disp  hp drat   wt qsec vs am gear carb
#> Merc 450SE 16.4   8 275.8 180 3.07 4.07 17.4  0  0    3    3
#> 
#> [[13]]
#>             mpg cyl  disp  hp drat   wt qsec vs am gear carb
#> Merc 450SL 17.3   8 275.8 180 3.07 3.73 17.6  0  0    3    3
#> 
#> [[14]]
#>              mpg cyl  disp  hp drat   wt qsec vs am gear carb
#> Merc 450SLC 15.2   8 275.8 180 3.07 3.78   18  0  0    3    3
#> 
#> [[15]]
#>                     mpg cyl disp  hp drat   wt  qsec vs am gear carb
#> Cadillac Fleetwood 10.4   8  472 205 2.93 5.25 17.98  0  0    3    4
#> 
#> [[16]]
#>                      mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Lincoln Continental 10.4   8  460 215    3 5.424 17.82  0  0    3    4
#> 
#> [[17]]
#>                    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Chrysler Imperial 14.7   8  440 230 3.23 5.345 17.42  0  0    3    4
#> 
#> [[18]]
#>           mpg cyl disp hp drat  wt  qsec vs am gear carb
#> Fiat 128 32.4   4 78.7 66 4.08 2.2 19.47  1  1    4    1
#> 
#> [[19]]
#>              mpg cyl disp hp drat    wt  qsec vs am gear carb
#> Honda Civic 30.4   4 75.7 52 4.93 1.615 18.52  1  1    4    2
#> 
#> [[20]]
#>                 mpg cyl disp hp drat    wt qsec vs am gear carb
#> Toyota Corolla 33.9   4 71.1 65 4.22 1.835 19.9  1  1    4    1
#> 
#> [[21]]
#>                mpg cyl  disp hp drat    wt  qsec vs am gear carb
#> Toyota Corona 21.5   4 120.1 97  3.7 2.465 20.01  1  0    3    1
#> 
#> [[22]]
#>                   mpg cyl disp  hp drat   wt  qsec vs am gear carb
#> Dodge Challenger 15.5   8  318 150 2.76 3.52 16.87  0  0    3    2
#> 
#> [[23]]
#>              mpg cyl disp  hp drat    wt qsec vs am gear carb
#> AMC Javelin 15.2   8  304 150 3.15 3.435 17.3  0  0    3    2
#> 
#> [[24]]
#>             mpg cyl disp  hp drat   wt  qsec vs am gear carb
#> Camaro Z28 13.3   8  350 245 3.73 3.84 15.41  0  0    3    4
#> 
#> [[25]]
#>                   mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Pontiac Firebird 19.2   8  400 175 3.08 3.845 17.05  0  0    3    2
#> 
#> [[26]]
#>            mpg cyl disp hp drat    wt qsec vs am gear carb
#> Fiat X1-9 27.3   4   79 66 4.08 1.935 18.9  1  1    4    1
#> 
#> [[27]]
#>               mpg cyl  disp hp drat   wt qsec vs am gear carb
#> Porsche 914-2  26   4 120.3 91 4.43 2.14 16.7  0  1    5    2
#> 
#> [[28]]
#>               mpg cyl disp  hp drat    wt qsec vs am gear carb
#> Lotus Europa 30.4   4 95.1 113 3.77 1.513 16.9  1  1    5    2
#> 
#> [[29]]
#>                 mpg cyl disp  hp drat   wt qsec vs am gear carb
#> Ford Pantera L 15.8   8  351 264 4.22 3.17 14.5  0  1    5    4
#> 
#> [[30]]
#>               mpg cyl disp  hp drat   wt qsec vs am gear carb
#> Ferrari Dino 19.7   6  145 175 3.62 2.77 15.5  0  1    5    6
#> 
#> [[31]]
#>               mpg cyl disp  hp drat   wt qsec vs am gear carb
#> Maserati Bora  15   8  301 335 3.54 3.57 14.6  0  1    5    8
#> 
#> [[32]]
#>             mpg cyl disp  hp drat   wt qsec vs am gear carb
#> Volvo 142E 21.4   4  121 109 4.11 2.78 18.6  1  1    4    2
#>